requirements.txt
pandas~=2.2.3
Faker~=30.8.0
psycopg2~=2.9.10
tabulate~=0.9.0
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1